Remove canonical tags for now
This is interfering with analytics. When the `routeChangeComplete`
event fires, the contents of the Head has not changed yet (or
at least is not guaranteed to have changed). Fathom uses the
canonical tag (if it's present) to track the page view, which is
resulting in tracking a view for the previous page instead of the
current one.